Manager dao , service ,, controller and vieww >all updated / created

+++ SmartStudent/src/main/java/com/manage/manager/service/Manager_FeeServiceImpl.java
@@ -1,46 +1,68 @@
 package com.manage.manager.service;
 
-import com.manage.student.entities.Fee;
 import com.manage.manager.dao.Manager_FeeDAO;
+import com.manage.student.entities.Fee;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
-@Transactional("managerTransactionManager")
+@Qualifier("managerFeeServiceImpl")
 public class Manager_FeeServiceImpl implements Manager_FeeService {
 
     private final Manager_FeeDAO managerFeeDAO;
 
     @Autowired
-    public Manager_FeeServiceImpl(Manager_FeeDAO managerFeeDAO) {
+    public Manager_FeeServiceImpl(@Qualifier("managerFeeDAOImpl") Manager_FeeDAO managerFeeDAO) {
         this.managerFeeDAO = managerFeeDAO;
     }
 
     @Override
-    public void saveFee(Fee fee) {
-        managerFeeDAO.saveFee(fee);
-    }
-
-    @Override
-    public void updateFee(Fee fee) {
-        managerFeeDAO.updateFee(fee);
-    }
-
-    @Override
-    public void deleteFee(Long feeId) {
-        managerFeeDAO.deleteFee(feeId);
-    }
-
-    @Override
+    @Transactional(transactionManager = "managerTransactionManager")
     public Fee getFeeById(Long feeId) {
         return managerFeeDAO.getFeeById(feeId);
     }
 
     @Override
+    @Transactional(transactionManager = "managerTransactionManager")
+    public void saveFee(Fee fee) {
+        managerFeeDAO.saveFee(fee);
+    }
+
+    @Override
+    @Transactional(transactionManager = "managerTransactionManager")
+    public void updateFee(Fee fee) {
+        managerFeeDAO.updateFee(fee);
+    }
+
+    @Override
+    @Transactional(transactionManager = "managerTransactionManager")
+    public void deleteFee(Long feeId) {
+        Fee fee = getFeeById(feeId);
+        if (fee != null) {
+            managerFeeDAO.deleteFee(feeId);
+        }
+    }
+
+    @Override
+    @Transactional(transactionManager = "managerTransactionManager")
     public List<Fee> getAllFees() {
         return managerFeeDAO.getAllFees();
+    }
+
+    @Override
+    @Transactional(transactionManager = "managerTransactionManager")
+    public List<Fee> getFeesByStudentId(Long studentId) {
+        return managerFeeDAO.getFeesByStudentId(studentId);
+    }
+    
+    @Override
+    @Transactional(transactionManager = "managerTransactionManager")
+    public Long saveFeeAndReturnId(Fee fee) {
+        managerFeeDAO.saveFee(fee);
+        return fee.getFeeId();
     }
 }
